2010-07-28  Angelo Corsaro  <angelo.corsaro@gmail.com>
git-svn-id: http://simd-cxx.googlecode.com/svn/trunk@111 2a94243a-7533-11de-bf39-f1df4430aa60

+++ demo/ddj-series/01/util.cpp
@@ -16,7 +16,7 @@ const float              H0 = 0.5F;
 const float              DH = 0.2;
 
 timespec msec2timespec(unsigned long long msec) {
-  unsigned long long nsec = msec * 1E6;
+  unsigned long int nsec = msec * 1E6;
   timespec ts = {0, nsec};
   // nsec has to be <= 999999999
